Сделал более аккуратнее и логичнее
